chore: add simple stop command
The child process handle is stored in the workspace state (might not be
so safe between editor restarts), so when run with `kamel --dev` sending
`CTRL+C` (`SIGINT`) will stop/delete the running kamel integration.

Otherwise if the kamel integration is run outside of the workspace it
will be stopped/deleted via `kamel delete`.

Super POCy :)