FUSETOOLS2-224 - provide template for Camel K in tasks.json
- single static template for starting
- using CompletionItemprovider, it is the same API than for a Language
Server, so the day we want to extract it as a more generic tool it
should be relatively easy
- using a jsonc-parser. it is the parser used by
vscode-json-languageservice. The VS Code JSOn Language Service is not
proposing useful API to reuse directly.
- the computation of the place where to provide the completion is a bit
handcrafted, would be best to rely on json schema in the future.
Currently, it will provide false-positive only if a tasks.json is used
and is not a VS Code Tasks files and that inside there is a tasks
attribute with an Array. I think the risk of false positive is low
enough for a first iteration.

Signed-off-by: Aurélien Pupier <apupier@redhat.com>