FUSETOOLS2-259 - Skip completion test on Jenkins
we already deactivated the test for completion for Java files when
running on Jenkins for VS Code Camel K files.
I guess the easiest for now is deactivate this one too as it is playing
correctly locally and on Travis and already spent a fair amount of time
investigating without success.

Signed-off-by: Aurélien Pupier <apupier@redhat.com>