FUSETOOLS2-4 - improve message error when Camel K installation not correct
would be better if underlying kamel/kubectl provides more precise
message but I think they are targeting users with different set of
skills and it will require a lot more work to include it upstream. So as
a first iteration,
here is this not very nice but  improving slightly UX for users.

testing it would require quite a lot of refactor and mock. Not sure if
it is worthy.

Signed-off-by: Aurélien Pupier <apupier@redhat.com>