FUSETOOLS2-813 - support space to path of executable kamel and kubectl
nota: given that the current bad behavior is that binary on system path
takes precedence over manually specified tool path, the tests require to
not have set the kamel and kubectl on system path. It also requires to
be launched on Linux as binaries are provided for linux only fo rthe
test.

Signed-off-by: Aurélien Pupier <apupier@redhat.com>