FUSETOOLS2-878 - Update default path to Camel K schema
we were pointing to master as it was implemented when there was no tag
available yet. it was refactored on master. So pointing to latest tag.
It will avoid breakage from external part but requires more maintenance.

including more traces in log too for next time, it will be easier to
debug

Signed-off-by: Aurélien Pupier <apupier@redhat.com>