FUSETOOLS2-944 - provide specific task to activate Camel K debug
this is a first iteration which avoids to go back to command line for
most cases. Even if it requires several config to be setup first. It
also has the benefit to allow to investigate possibilities more easily
for better support.

- it might allow to combine with a Java Debug
launch config
- the parameter for integrationName might need to be different and take
a file name and convert? or both?
- might be better to have a specific launch config instead of task (but
sounds far more complicated)

Signed-off-by: Aurélien Pupier <apupier@redhat.com>