FUSETOOLS2-1128 - Provide Start CodeLens on Camel K files
* available for all types of files with heuristic:
** have *.camelk.* in file name
** or have Camel K modeline
* currently triggers the command to "Start Integration", it lets up to
the user to choose which kind of start he wants. In the future, we might
decide to pick basic, or dev or debug or an associated task or the
latest one used? But it is the easiest to implement as a first iteration
and it already provides a lot of improvement
* there is no test executing the codelenses. I have not found a way to
do it programmatically through VS Code API. I reported this issue to
provide a UI test https://issues.redhat.com/browse/FUSETOOLS2-1213
* Readme and tutorials remains valid to start an integration. That said,
it might be nice to update them to include the usage of the Codelens.
Reported https://issues.redhat.com/browse/FUSETOOLS2-1214 to discuss it
and potentially do it in another iteration. For now, simply added a note
as a minimal change.

Signed-off-by: Aurélien Pupier <apupier@redhat.com>