FUSETOOLS2-1799 - replace download package
- download package is no more maintained and contains a security
vulnerability through its chain of dependency
- using node-fetch 2.x despite 3.x is available. Unfortunately, it is
not working as-is due to ESM module. 2.x is maintained for security fix
and important bugs
- lost the functionality which was sharing progress of download. i have
not found a maintained library capable of it and wasn't able to
implement it with node-fetch and tar packages. I think this can be done
in another iteration anyway. It is more important to have a secure
extension.

Signed-off-by: Aurélien Pupier <apupier@redhat.com>