Fix refresh of classpath for some MacOS
For not clearly defined reasons (that we are able to find), the current
working directory used by default by Node.js execSync method is not
picked by correctly by Camel Jbang. Using the workspace folder is
workarounding the issue.
In regular cases, there will be a workspace folder available.

fix #1698

Signed-off-by: Aurélien Pupier <apupier@redhat.com>