Test with 1.85.2 instead of latest 1.86
it is failing with 1.86.x. Cannot reproduce manually even if behavior
seems inconsistent, it is always possible to refresh classpath several
times and/or close/reopen and it finishes to work.

To unblock main branch, test with 1.85 and add a note in readme.

Reported https://issues.redhat.com/browse/FUSETOOLS2-2311 to tackle it
in another phase.

Signed-off-by: Aurélien Pupier <apupier@redhat.com>